modify code about the navbar
