+++ src/components/nav-bar/nav-bar.jsx
@@ -1,0 +1,26 @@
+import { Link, useLocation } from "react-router-dom";
+
+export default function NavBar() {
+    const excluded_routes = ["/details", "/profile-update"];
+
+    const location = useLocation();
+
+    if (excluded_routes.includes(location.pathname)) return null;
+
+    return (
+        <nav>
+          <ul>
+            <li>
+              {/* due to the error "No routes matched location '/'" now making the registration page to take the space ie acting like the index page */}
+              <Link to="/">Registration</Link>
+            </li>
+            <li>
+              <Link to="/details">Details</Link>
+            </li>
+            <li>
+              <Link to="/profile-update">Profile Update</Link>
+            </li>
+          </ul>
+        </nav>
+    )
+}
