+++ src/pages/details/details.module.css
