+++ src/pages/details/index.jsx
