+++ src/pages/profileUpdate/index.jsx
@@ -86,6 +86,7 @@
 
 // export default ProfileUpdatePage;
 
+// eslint-disable-next-line no-unused-vars
 import React, { useEffect, useState } from 'react';
 import styles from './update.module.css';
