+++ src/pages/profileUpdate/update.module.css
@@ -12,13 +12,13 @@
 }
 
 .form {
-  width: 60vw;
+  width: 80vw;
   display: flex;
   flex-direction: column;
 }
 
 .formGroup {
-  max-width: 40%;
+  max-width: 80%;
   margin: 16px auto;
   justify-content: center;
   justify-items: center;
