+++ src/pages/registration/index.jsx
@@ -1,0 +1,123 @@
+// eslint-disable-next-line no-unused-vars
+import React from "react";
+import { useFormik } from "formik";
+import "./registration.module.css";
+import * as Yup from "yup";
+import { useHistory } from 'react-router-dom';
+
+const RegistrationForm = () => {
+  const history = useHistory();
+
+  const formik = useFormik({
+    initialValues: {
+      firstName: "",
+      lastName: "",
+      email: "",
+      phoneNumber: "",
+    },
+    validationSchema: Yup.object({
+      firstName: Yup.string().required("First Name is required"),
+      lastName: Yup.string().required("Last Name is required"),
+      email: Yup.string()
+        .email("Invalid email address")
+        .required("Email is required"),
+      phoneNumber: Yup.string()
+        .matches(/^[6-9]\d{8}$/, "Invalid phone number")
+        .required("Phone Number is required"),
+    }),
+
+    onSubmit: (values) => {
+      // Handle form submission
+      console.log(values);
+
+      // Store the form data in localStorage
+      const formData = {
+        firstName: values.firstName,
+        lastName: values.lastName,
+        email: values.email,
+        phoneNumber: values.phoneNumber,
+      };
+
+      localStorage.setItem("formData", JSON.stringify(formData));
+
+      // Reset the form fields
+      formik.resetForm();
+
+      // Redirect to the details page
+      history.push('/details');
+    },
+  });
+
+  return (
+    <>
+      <h1 className="heading">Registration</h1>
+      <form onSubmit={formik.handleSubmit}>
+        <div>
+          <label htmlFor="firstName">First Name:</label>
+          <input
+            type="text"
+            id="firstName"
+            name="firstName"
+            onChange={formik.handleChange}
+            onBlur={formik.handleBlur}
+            value={formik.values.firstName}
+          />
+          {formik.touched.firstName && formik.errors.firstName && (
+            <div>{formik.errors.firstName}</div>
+          )}
+        </div>
+
+        <div>
+          <label htmlFor="lastName">Last Name:</label>
+          <input
+            type="text"
+            id="lastName"
+            name="lastName"
+            onChange={formik.handleChange}
+            onBlur={formik.handleBlur}
+            value={formik.values.lastName}
+          />
+          {formik.touched.lastName && formik.errors.lastName && (
+            <div>{formik.errors.lastName}</div>
+          )}
+        </div>
+
+        <div>
+          <label htmlFor="email">Email:</label>
+          <input
+            type="email"
+            id="email"
+            name="email"
+            onChange={formik.handleChange}
+            onBlur={formik.handleBlur}
+            value={formik.values.email}
+          />
+          {formik.touched.email && formik.errors.email && (
+            <div>{formik.errors.email}</div>
+          )}
+        </div>
+
+        <div>
+          <label htmlFor="phoneNumber">Phone Number:</label>
+          <input
+            type="text"
+            id="phoneNumber"
+            name="phoneNumber"
+            onChange={formik.handleChange}
+            onBlur={formik.handleBlur}
+            value={formik.values.phoneNumber}
+          />
+          {formik.touched.phoneNumber && formik.errors.phoneNumber && (
+            <div>{formik.errors.phoneNumber}</div>
+          )}
+        </div>
+
+        <button type="submit" className="button1">
+          Create Profile
+        </button>
+      </form>
+    </>
+  );
+};
+
+export default RegistrationForm;
